seperate spdz code into separate file
